III-4475 - Show PriceInformation directly in tab content instead of modal
